Rename actions to action
