+++ Typography/src/index.ts
@@ -1,7 +1,7 @@
 export { default as Header } from './header/Header';
 export { default as Title } from './title/Title';
 export { default as Paragraph } from './paragraph/Paragraph';
-export { default as Actions } from './actions/Actions';
+export { default as Action } from './action/Action';
 export { default as Label } from './label/Label';
 export { default as Helper } from './helper/Helper';
 export { default as Numeric } from './numeric/Numeric';
